+++ lib/src/scope/h_scope/scope_base.dart
@@ -57,6 +57,13 @@ abstract base class Scope<W extends Scope<W, D, S>, D extends ScopeDependencies,
   /// A callback invoked when the [scopeKeyTimeout] expires.
   final void Function()? onScopeKeyTimeout;
 
+  /// The timeout duration for cancelling the initialization during the
+  /// teardown, before it triggers [onInitCancellationTimeout].
+  final Duration? initCancellationTimeout;
+
+  /// A callback invoked when the [initCancellationTimeout] expires.
+  final void Function()? onInitCancellationTimeout;
+
   /// The timeout duration for waiting to dispose child scopes.
   final Duration? waitForChildrenTimeout;
 
@@ -73,6 +80,8 @@ abstract base class Scope<W extends Scope<W, D, S>, D extends ScopeDependencies,
     this.scopeKey,
     this.scopeKeyTimeout,
     this.onScopeKeyTimeout,
+    this.initCancellationTimeout,
+    this.onInitCancellationTimeout,
     this.waitForChildrenTimeout,
     this.onWaitForChildrenTimeout,
     this.pauseAfterInitialization,
@@ -190,6 +199,12 @@ final class _ScopeElement<W extends Scope<W, D, S>, D extends ScopeDependencies,
 
   @override
   void onScopeKeyTimeout() => widget.onScopeKeyTimeout?.call();
+
+  @override
+  Duration? get initCancellationTimeout => widget.initCancellationTimeout;
+
+  @override
+  void onInitCancellationTimeout() => widget.onInitCancellationTimeout?.call();
 
   @override
   Duration? get waitForChildrenTimeout => widget.waitForChildrenTimeout;